+++ packages/libredwg-converter/src/AcDbLibreDwgConverter.ts
@@ -522,6 +522,13 @@ export class AcDbLibreDwgConverter extends AcDbDatabaseConverter<DwgDatabase> {
           break
         }
       }
+      // If the layout is not found in the block table due to some unknow reason, 
+      // let's set the model space block table record id.
+      if (!dbLayout.blockTableRecordId) {
+        if (layout.layoutName === 'Model') {
+          dbLayout.blockTableRecordId = db.tables.blockTable.modelSpace.objectId
+        }
+      }
       dbLayout.limits.min.copy(layout.minLimit)
       dbLayout.limits.max.copy(layout.maxLimit)
       dbLayout.extents.min.copy(layout.minExtent)